clean up + folder rename
