+++ src/pages/backoffice/content/Content.js
@@ -2,15 +2,13 @@ import React from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faArrowRight } from "@fortawesome/free-solid-svg-icons";
 import { EditOrganization } from "../EditOrganization/EditOrganization";
-//import ListScreen from "../../backoffice/ListScreen/ListScreen.jsx";
 import { ListActivities } from "../ListActivities/ListActivities";
 import { EditHomeData } from "../EditHomeData/EditHomeData";
 import { Switch, Route } from "react-router-dom";
 import { ListUsers } from "../ListUsers/ListUsers";
 import { ListContact } from "../ListContact/ListContact";
 import CategoriesScreen from "../categories/CategoriesScreen";
-import EditActivities from "../../../components/EditActivities/EditActivities";
-import NoveltiesScreen from "../novedades/NoveltiesScreen";
+import NoveltiesScreen from "../novelties/NoveltiesScreen";
 
 export const Content = ({ openAside, isLeft }) => {
   const left = isLeft ? "icon isleft" : "icon isright";
