+++ src/pages/backoffice/novelties/NoveltiesList.jsx
