+++ src/pages/backoffice/novelties/novelties.scss
